pyproject.toml
[tool.poetry]
name = "resume"
version = "0.1.0"
description = ""
authors = ["MatthieuLeNozach <matthieu.le.nozach@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
jinja2 = "^3.1.4"
fastapi-cli = "^0.0.5"
pytailwindcss = "^0.2.0"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
